+++ kpresenter/kppieobject.cc
@@ -65,7 +65,7 @@ KPPieObject &KPPieObject::operator=( const KPPieObject & )
 }
 
 /*========================= save =================================*/
-QDomDocumentFragment KPPieObject::save( QDomDocument& doc, int offset )
+QDomDocumentFragment KPPieObject::save( QDomDocument& doc, double offset )
 {
   QDomDocumentFragment fragment=KP2DObject::save(doc, offset);
   if (lineBegin!=L_NORMAL)
@@ -82,9 +82,9 @@ QDomDocumentFragment KPPieObject::save( QDomDocument& doc, int offset )
 }
 
 /*========================== load ================================*/
-int KPPieObject::load(const QDomElement &element)
+double KPPieObject::load(const QDomElement &element)
 {
-    int offset=KP2DObject::load(element);
+    double offset=KP2DObject::load(element);
     QDomElement e=element.namedItem("LINEBEGIN").toElement();
     if(!e.isNull()) {
         int tmp=0;
