+++ kpresenter/kppieobject.h
@@ -80,8 +80,8 @@ public:
     virtual LineEnd getLineEnd() const
     { return lineEnd; }
 
-    virtual QDomDocumentFragment save( QDomDocument& doc, int offset );
-    virtual int load(const QDomElement &element);
+    virtual QDomDocumentFragment save( QDomDocument& doc, double offset );
+    virtual double load(const QDomElement &element);
 
 protected:
     virtual void paint( QPainter *_painter, KoZoomHandler*_zoomHandler, bool drawingShadow );
