+++ kpresenter/kppolylineobject.cc
@@ -59,7 +59,7 @@ KPPolylineObject &KPPolylineObject::operator=( const KPPolylineObject & )
 }
 
 /*========================= save =================================*/
-QDomDocumentFragment KPPolylineObject::save( QDomDocument& doc, int offset )
+QDomDocumentFragment KPPolylineObject::save( QDomDocument& doc, double offset )
 {
     QDomDocumentFragment fragment = KPObject::save( doc, offset );
     fragment.appendChild( KPObject::createPenElement( "PEN", pen, doc ) );
@@ -87,9 +87,9 @@ QDomDocumentFragment KPPolylineObject::save( QDomDocument& doc, int offset )
 }
 
 /*========================== load ================================*/
-int KPPolylineObject::load(const QDomElement &element)
+double KPPolylineObject::load(const QDomElement &element)
 {
-    int offset=KPObject::load( element );
+    double offset=KPObject::load( element );
     QDomElement e = element.namedItem( "PEN" ).toElement();
     if ( !e.isNull() )
         setPen( KPObject::toPen( e ) );
